feat: add caching, YOLOv8 detector, IRI ingestion, address search, and setup guide
- Add in-memory TTL caching for segments (5min) and routes (2min)
  with /cache/stats and /cache/clear endpoints
- Add YOLOv8 pothole detector with factory fallback to stub
- Add IRI ingestion scripts (CSV, shapefile, improved synthetic)
- Add AddressInput component with Nominatim autocomplete for Route Finder
- Add comprehensive setup/run/usage guide (docs/SETUP.md)
- Add integration, cache, YOLOv8, and IRI ingestion tests
- Remove tracked __pycache__ files, update .gitignore

+++ db/migrations/001_initial.sql
@@ -4,8 +4,8 @@ CREATE TABLE IF NOT EXISTS road_segments (
     geom          GEOMETRY(LineString, 4326) NOT NULL,
     length_m      DOUBLE PRECISION NOT NULL,
     travel_time_s DOUBLE PRECISION NOT NULL,
-    source        INTEGER,
-    target        INTEGER,
+    source        BIGINT,
+    target        BIGINT,
     iri_value     DOUBLE PRECISION,
     iri_norm      DOUBLE PRECISION,
     created_at    TIMESTAMPTZ DEFAULT NOW()
